Complete refactor of project structure.
