+++ pyatlas/doc/how_to_get_the_docs.txt
@@ -1,2 +1,6 @@
-Run the 'buildPyatlas' target to generate the docs!
+Use the 'buildPyatlas' gradle target to generate the docs!
+
+From the root atlas project directory run:
+
+$ ./gradlew buildPyatlas
