+++ pyatlas/unit_tests/test_polygon_converters.py
@@ -1,9 +1,9 @@
 import unittest
 
-from pyatlas.location import Location
-from pyatlas.polygon import Polygon
-from pyatlas.polyline import PolyLine
-from pyatlas import polygon_converters
+from pyatlas import geometry
+from pyatlas.geometry import Location
+from pyatlas.geometry import Polygon
+from pyatlas.geometry import PolyLine
 import shapely.geometry
 
 
@@ -21,7 +21,7 @@ class PolygonConvertersTest(unittest.TestCase):
         ]
         bounds = Polygon(loclist).get_bounds()
 
-        shapely_box = polygon_converters.boundable_to_shapely_box(bounds)
+        shapely_box = geometry.boundable_to_shapely_box(bounds)
         test_against = shapely.geometry.LineString([(0, 0), (450000000, 0), (450000000, 450000000),
                                                     (0, 450000000)])
         test_against = shapely.geometry.Polygon(test_against)
@@ -38,7 +38,7 @@ class PolygonConvertersTest(unittest.TestCase):
         ]
         polygon = Polygon(loclist)
 
-        shapely_poly = polygon_converters.polygon_to_shapely_polygon(polygon)
+        shapely_poly = geometry.polygon_to_shapely_polygon(polygon)
         test_against = shapely.geometry.LineString([(0, 0), (400000000, 0), (350000000, 300000000),
                                                     (450000000, 450000000), (1000, 450000000)])
         test_against = shapely.geometry.Polygon(test_against)
@@ -50,9 +50,9 @@ class PolygonConvertersTest(unittest.TestCase):
         l2 = Location(1000, 2000)
         l3 = Location(50000, -1000000)
 
-        p1 = polygon_converters.location_to_shapely_point(l1)
-        p2 = polygon_converters.location_to_shapely_point(l2)
-        p3 = polygon_converters.location_to_shapely_point(l3)
+        p1 = geometry.location_to_shapely_point(l1)
+        p2 = geometry.location_to_shapely_point(l2)
+        p3 = geometry.location_to_shapely_point(l3)
 
         self.assertEquals(shapely.geometry.Point(0, 0), p1)
         self.assertEquals(shapely.geometry.Point(1000, 2000), p2)
@@ -60,6 +60,6 @@ class PolygonConvertersTest(unittest.TestCase):
 
     def test_polyline_to_shapely_linestring(self):
         polyline1 = PolyLine([Location(-1000, -1000), Location(0, 0), Location(5000, 8000)])
-        linestring1 = polygon_converters.polyline_to_shapely_linestring(polyline1)
+        linestring1 = geometry.polyline_to_shapely_linestring(polyline1)
         test_against = shapely.geometry.LineString([(-1000, -1000), (0, 0), (5000, 8000)])
         self.assertEquals(linestring1, test_against)
